split spawn test into different files
