+++ test/spawn/exit-code.js
@@ -1,0 +1,10 @@
+const tap = require('tap');
+
+const spawn = require('../utils/spawn');
+
+tap.test('should pass through the exit code', t => {
+  spawn('exit.js').on('exit', code => {
+    t.equal(code, 101);
+    t.end();
+  });
+});
