+++ test/spawn/index.js
@@ -1,169 +1,23 @@
-const semver = require('semver');
 const tap = require('tap');
 
 const spawn = require('../utils/spawn');
 const touchFile = require('../utils/touch-file');
 
-tap.test('should pass unknown args to node binary', t => {
-  spawn('--expose_gc gc.js foo', out => {
-    t.equal(out.trim(), 'foo function');
-    return { exit: t.end.bind(t) };
-  });
-});
-
-tap.test('should restart the server twice', t => {
-  spawn('server.js', out => {
-    if (out.match(/touch message.js/)) {
-      touchFile('message.js');
-      return out2 => {
-        if (out2.match(/Restarting/)) {
-          touchFile('message.js');
-          return out3 => {
-            if (out3.match(/Restarting/)) {
-              return { exit: t.end.bind(t) };
-            }
-          };
-        }
-      };
-    }
-  });
-});
-
-tap.test('should handle errors', t => {
-  spawn('error.js', out => {
-    if (out.match(/ERROR/)) {
-      touchFile('message.js');
-      return out2 => {
-        if (out2.match(/Restarting/)) {
-          return { exit: t.end.bind(t) };
-        }
-      };
-    }
-  });
-});
-
-tap.test('should handle null errors', t => {
-  spawn('error-null.js', out => {
-    if (out.match(/ERROR/)) {
-      touchFile('message.js');
-      return out2 => {
-        if (out2.match(/Restarting/)) {
-          return { exit: t.end.bind(t) };
-        }
-      };
-    }
-  });
-});
-
-tap.test('should watch if no such module', t => {
-  let passed = false;
-  spawn('no-such-module.js', out => {
-    if (!passed && out.match(/ERROR/)) {
-      passed = true;
-      return { exit: t.end.bind(t) };
-    }
-  });
-});
-
-tap.test('should run async code un uncaughtException handlers', t => {
-  spawn('uncaught-exception-handler.js', out => {
-    if (out.match(/ERROR/)) {
-      return out2 => {
-        if (out2.match(/async \[?ReferenceError/)) {
-          return { exit: t.end.bind(t) };
-        }
-      };
-    }
-  });
-});
-
-tap.test('should ignore caught errors', t => {
-  spawn('catch-no-such-module.js', out => {
-    t.match(out, /Caught/);
-    return { exit: t.end.bind(t) };
-  });
-});
-
-tap.test('should not show up in argv', t => {
-  spawn('argv.js foo', out => {
-    const argv = JSON.parse(out.replace(/'/g, '"'));
-    t.match(argv[0], /.*?node(js|\.exe)?$/);
-    t.equal(argv[1], 'argv.js');
-    t.equal(argv[2], 'foo');
-    return { exit: t.end.bind(t) };
-  });
-});
-
-tap.test('should pass through the exit code', t => {
-  spawn('exit.js').on('exit', code => {
-    t.equal(code, 101);
-    t.end();
-  });
-});
-
-tap.test('should conceal the wrapper', t => {
-  // require.main should be main.js not wrap.js!
-  spawn('main.js').on('exit', code => {
-    t.equal(code, 0);
-    t.end();
-  });
-});
-
-tap.test('should relay stdin', t => {
-  const p = spawn('echo.js', out => {
-    t.equal(out, 'foo');
-    return { exit: t.end.bind(t) };
-  });
-  p.stdin.write('foo');
-});
-
-tap.test('Restart the cluster', t => {
-  spawn('cluster.js', out => {
-    if (out.match(/touch message\.js/m)) {
-      touchFile('message.js');
-      return out2 => {
-        if (out2.match(/Restarting/m)) {
-          return out3 => {
-            if (out3.match(/All workers disconnected/m)) {
-              let shuttingDown = false;
-              return out4 => {
-                if (out4.match(/touch message\.js/) && !shuttingDown) {
-                  shuttingDown = true;
-                  return { exit: t.end.bind(t) };
-                }
-              };
-            }
-          };
-        }
-      };
-    }
-  });
-});
-
-tap.test('should kill the forked processes', t => {
-  spawn('pid.js', out => {
-    const pid = parseInt(out, 10);
-    return {
-      exit: () => {
-        setTimeout(() => {
-          try {
-            process.kill(pid);
-            t.fail('child must no longer run');
-          } catch (e) {
-            t.end();
-          }
-        }, 500);
-      }
-    };
-  });
-});
-
-tap.test('should *not* set NODE_ENV', t => {
-  spawn('env.js', out => {
-    t.notMatch(out, /development/);
-    return { exit: t.end.bind(t) };
-  });
-});
+require('./argv');
+require('./caught');
+require('./clear');
+require('./cluster');
+require('./conceal');
+require('./errors');
+require('./esmodule');
+require('./exit-code');
+require('./expose-gc');
+require('./kill-fork');
+require('./no-such-module');
+require('./node-env');
+require('./relay-stdin');
+require('./restart-twice');
+require('./uncaught');
 
 tap.test('should allow graceful shutdowns', t => {
   if (process.platform === 'win32') {
@@ -247,36 +101,6 @@ tap.test('Uses ts-node/register for .ts files through config file (also the defa
   });
 });
 
-tap.test('Supports ECMAScript modules', t => {
-  spawn('ecma-script-modules.mjs', out => {
-    if (out.match(/touch message.mjs/)) {
-      touchFile('message.mjs');
-      return out2 => {
-        if (out2.match(/Restarting/)) {
-          t.match(out2, /\[INFO\] \d{2}:\d{2}:\d{2} Restarting/);
-          return { exit: t.end.bind(t) };
-        }
-      };
-    }
-  });
-});
-
-tap.test('Supports ECMAScript modules with experimental-specifier-resolution', t => {
-  if (semver.satisfies(process.version, '<12.17')) return t.skip();
-
-  spawn('--experimental-specifier-resolution=node resolution.mjs', out => {
-    if (out.match(/touch message.js/)) {
-      touchFile('message.js');
-      return out2 => {
-        if (out2.match(/Restarting/)) {
-          t.match(out2, /\[INFO\] \d{2}:\d{2}:\d{2} Restarting/);
-          return { exit: t.end.bind(t) };
-        }
-      };
-    }
-  });
-});
-
 tap.test('Supports --inspect', t => {
   spawn('--inspect server.js', out => {
     if (out.match(/Debugger listening on/)) {
@@ -293,5 +117,3 @@ tap.test('Supports --inspect', t => {
     }
   });
 });
-
-require('./clear.js');
