+++ test/spawn/kill-fork.js
@@ -1,0 +1,21 @@
+const tap = require('tap');
+
+const spawn = require('../utils/spawn');
+
+tap.test('should kill the forked processes', t => {
+  spawn('pid.js', out => {
+    const pid = parseInt(out, 10);
+    return {
+      exit: () => {
+        setTimeout(() => {
+          try {
+            process.kill(pid);
+            t.fail('child must no longer run');
+          } catch (e) {
+            t.end();
+          }
+        }, 500);
+      }
+    };
+  });
+});
