+++ test/spawn/node-env.js
@@ -1,0 +1,10 @@
+const tap = require('tap');
+
+const spawn = require('../utils/spawn');
+
+tap.test('should *not* set NODE_ENV', t => {
+  spawn('env.js', out => {
+    t.notMatch(out, /development/);
+    return { exit: t.end.bind(t) };
+  });
+});
