+++ test/spawn/relay-stdin.js
@@ -1,0 +1,10 @@
+const tap = require('tap');
+
+const spawn = require('../utils/spawn');
+
+tap.test('should relay stdin', t => {
+  spawn('echo.js', out => {
+    t.equal(out, 'foo');
+    return { exit: t.end.bind(t) };
+  }).stdin.write('foo');
+});
